refactor: move SizeUnit and FontSizeValue into ast/font_size_value.hh
Extract ::pltxt2htm::SizeUnit and ::pltxt2htm::FontSizeValue from
ast/node_kind.hh into a dedicated public header so node_kind.hh only
holds node-kind declarations. node_kind.hh re-includes the new header
to keep existing transitive and external consumers working, and direct
consumers (parser, optimizer, frame_context, try_parse, md_table, AST
impl decls, backends) now include it explicitly.

Also pass HtmlSpan's optional font-size by value instead of by rvalue
reference, matching PlSize: optional<FontSizeValue> is trivially
copyable, so && adds no benefit.
