+++ include/pltxt2htm/ast/impl/markdown_node_decl.hh
@@ -13,6 +13,7 @@
 #include <fast_io/fast_io_dsal/string.h>
 #include "ast_decl.hh"
 #include "basic_node_decl.hh"
+#include "../font_size_value.hh"
 #include "../node_kind.hh"
 
 namespace pltxt2htm {
