+++ include/pltxt2htm/ast/node_kind.hh
@@ -8,8 +8,6 @@
 
 #pragma once
 
-#include <cstddef>
-
 namespace pltxt2htm {
 
 /**
@@ -244,30 +242,6 @@ enum class TableAlign : unsigned {
     left = 0,
     center,
     right,
-};
-
-/**
- * @brief Unit of a font-size value
- * @details Distinguishes absolute sizes from percentages in &lt;size=N&gt; /
- *          &lt;size=N%&gt; (TMP_Text rich text) and &lt;span style="font-size:..."&gt;.
- */
-enum class SizeUnit : unsigned {
-    px = 0, ///< Absolute size (CSS px; TMP_Text unitless value)
-    percent, ///< Percentage of the default font size (TMP_Text / CSS %)
-};
-
-/**
- * @brief A font-size value together with its unit.
- * @details Pairs the numeric value with its ::pltxt2htm::SizeUnit so the two
- *          never travel independently. This avoids carrying an invalid/undefined
- *          unit when no font-size is present (the value is then simply absent).
- */
-struct FontSizeValue {
-    ::std::size_t value; ///< Numeric font-size value
-    ::pltxt2htm::SizeUnit unit; ///< Unit of the value (px or %)
-
-    [[nodiscard]]
-    constexpr auto operator==(::pltxt2htm::FontSizeValue const& other) const noexcept -> bool = default;
 };
 
 } // namespace pltxt2htm
