+++ include/pltxt2htm/details/backend/for_plweb_text.hh
@@ -13,6 +13,7 @@
 #include <fast_io/fast_io_dsal/string.h>
 #include <fast_io/fast_io_dsal/string_view.h>
 #include <exception/exception.hh>
+#include "../../ast/font_size_value.hh"
 #include "frame_context.hh"
 #include "../../contracts.hh"
 #include "../../details/utils.hh"
