+++ include/pltxt2htm/details/parser/try_parse.hh
@@ -14,6 +14,7 @@
 #include "../utils.hh"
 #include "../../contracts.hh"
 #include "../../ast/ast.hh"
+#include "../../ast/font_size_value.hh"
 #include "../push_macro.hh"
 
 /**
