+++ include/pltxt2htm/optimizer.hh
@@ -17,6 +17,7 @@
 #include <fast_io/fast_io_dsal/list.h>
 #include <fast_io/fast_io_dsal/stack.h>
 #include "ast/ast.hh"
+#include "ast/font_size_value.hh"
 #include "contracts.hh"
 #include "details/utils.hh"
 #include "details/push_macro.hh"
