+++ include/pltxt2htm/parser.hh
@@ -11,6 +11,7 @@
 #include <fast_io/fast_io_dsal/list.h>
 #include <fast_io/fast_io_dsal/stack.h>
 #include <fast_io/fast_io_dsal/string_view.h>
+#include "ast/font_size_value.hh"
 #include "ast/node_kind.hh"
 #include "contracts.hh"
 #include "details/utils.hh"
